+++ src/main/java/task/list/Event.java
@@ -1,4 +1,4 @@
-package list;
+package task.list;
 
 public class Event extends Deadline {
     private static final String TASK_TITLE = "E";
